Fixing warnings caused by catching exception by value

+++ tests/InputTest.cpp
@@ -37,7 +37,7 @@ void printGameplayVector(std::vector<std::string> arguments) {
     std::cout << "=== Vector Output ===" << std::endl;
     std::cout << "Arguments: " << arguments.size() << std::endl;
     std::cout << std::endl;
-    for(int i = 0; i < arguments.size(); i++) {
+    for(int i = 0; i < (int)arguments.size(); i++) {
         std::cout << i << ": " << arguments[i] << std::endl;
     }
     std::cout << std::endl;
